tests/wpt/tests/css/css-syntax/charset/page-windows-1251-css-at-charset-bogus-charset-attribute-windows-1250.html

<!doctype html>
<title>CSS charset: page windows-1251, CSS @charset bogus, charset attribute windows-1250</title>
<link rel=help href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding">
<meta charset=windows-1251>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel=stylesheet href="support/at-charset-bogus.css" charset=windows-1250>
<div id=log></div>
<div id=&#x010C;></div>
<script>
var elm = document.getElementById('\u010C');
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm, '').visibility, 'hidden');
  this.done();
});
</script>
